requirements.txt
eclipse-zenoh~=1.9.0
